DFI-1820 fixing coverage
